Installation method using plug-vim plugin manager
This commit changes the way vimtemplates is installed.

You can still run it from a manual instalation, as it recognizes both
directories to source templates.

But from now on, you just need to add to your plug-vim session the
lines:

" Vimtemplates - templates for diverse files
Plug 'drbeco/vimtemplates'

and voila! It will be downloaded (cloned from github) and installed,
and updated if a new version comes to be.

The new directory is:

~/.vim/plugged/vimtemplates/plugin/
and
~/.vim/plugged/vimtemplates/templates/